♻️ Réorganisation des fonctions de mode
