+++ src/core/commands/command_invoker.rs
@@ -1,9 +1,8 @@
-use crate::core::{editor::Editor, Core};
-
+use crate::core::Core;
 use super::Command;
 
 pub struct CommandInvoker<'a> {
-    pub core: &'a mut Core
+    pub core: &'a mut Core,
 }
 
 impl<'a> CommandInvoker<'a> {
@@ -13,7 +12,7 @@ impl<'a> CommandInvoker<'a> {
 
     pub fn execute_command(self: &mut CommandInvoker<'a>, mut command: impl Command + 'static) {
         command.execute_editor(&mut self.core.editor);
-        command.execute_core(&mut self.core);
+        command.execute_core(self.core);
         command.execute_cmd_line(&mut self.core.command_line);
     }
 }
