+++ src/core/commands/mod.rs
@@ -6,7 +6,7 @@ pub mod set_mode_command;
 pub mod write_cmd_line_command;
 
 pub trait Command {
-    fn execute_core(&mut self, core: &mut Core) {}
-    fn execute_editor(&mut self, editor: &mut Editor) {}
-    fn execute_cmd_line(&mut self, cmd: &mut CommandLine) {}
+    fn execute_core(&mut self, _core: &mut Core) {}
+    fn execute_editor(&mut self, _editor: &mut Editor) {}
+    fn execute_cmd_line(&mut self, _cmd: &mut CommandLine) {}
 }
