+++ src/core/commands/scroll_command.rs
@@ -1,19 +1,17 @@
 use std::io::BufRead;
-
 use crate::core::editor::Editor;
-
 use super::Command;
 
 pub struct ScrollUpCommand;
 pub struct ScrollDownCommand;
 
-impl Command for ScrollUpCommand  {
+impl Command for ScrollUpCommand {
     fn execute_editor(&mut self, editor: &mut Editor) {
         editor.scroll_up();
     }
 }
 
-impl Command for ScrollDownCommand  {
+impl Command for ScrollDownCommand {
     fn execute_editor(&mut self, editor: &mut Editor) {
         editor.scroll_down();
     }
@@ -23,7 +21,9 @@ impl Editor {
     /// Permet de scroller le text de [`Editor`](struct@Editor) d'un ligne vers le haut
     fn scroll_up(self: &mut Editor) {
         let num_lines = self.text_buffer.lines().count();
-        if num_lines == 0 { return; }
+        if num_lines == 0 {
+            return;
+        }
         if self.scroll.0 < num_lines as u16 - 1 {
             self.scroll.0 += 1;
         }
