+++ src/core/mod.rs
@@ -1,8 +1,3 @@
-/// Gestion de l'editeur **Cosmo**
-mod editor;
-/// Gestion des commandes
-pub mod commands;
-
 use command_line::CommandLine;
 use editor::Editor;
 use ratatui::{
@@ -11,12 +6,16 @@ use ratatui::{
     Terminal,
 };
 use std::io::Result;
-
 use crate::{input::handler::handler_input, ui::ui};
-pub mod command_line;
-pub mod editor_modes;
 
+/// Gestion de toutes les actions possible dans **Cosmo**
+pub mod commands;
+/// Gestion de l'editeur **Cosmo**
+mod editor;
+pub mod command_line;
+pub mod modes;
 mod utils;
+
 
 /// Liste les différents modes d'interaction de l'IDE.
 #[derive(PartialEq, Eq, Debug, Clone, Copy)]
@@ -27,14 +26,14 @@ pub enum InputMode {
     Exit,
 }
 
-/// Representation des donnees de **Cosmo**
+/// Representation de tous les modules de **Cosmo**
 pub struct Core {
     pub editor: Editor,
     pub command_line: CommandLine,
 }
 
 impl Core {
-    /// Crée une nouvelle instance de [`App`].
+    /// Crée une nouvelle instance de [`Core`].
     pub fn new() -> Self {
         Self {
             editor: Editor::new(),
