+++ src/core/modes/mod.rs
@@ -1,0 +1,3 @@
+pub mod normal_mode;
+pub mod insert_mode;
+pub mod command_mode;
