+++ src/input/mod.rs
@@ -1,1 +1,2 @@
+/// Recupere toutes les interactions au clavier
 pub mod handler;
